mmc: msm_sdcc: Fix the bug in platform resume method
If the config MMC_UNSAFE_RESUME is not defined, the suspend/resume is
basically a removal/insertion event.  The current code resumes the host
only if there is a card and is not SDIO type.  Resume the host even if
no card is present or if card is not SDIO type.

Sahitya Tummala <stummala@qualcomm.com>